Use AStar to compute travel paths, use travel instead of seek.
Previously, a DDA/Brasenham algorithm was used to compute paths between
NPCs and the player when NPCs were giving chase.  However, since a
different algorithm was being used to determine light of sight, (via ROT)
NPCs could often see the player without being able to determine a valid
path.  Using AStar allows NPCs to follow a non-linear path in the sense
that the path between them and player does not need to be a straight line.

Signed-off-by: suicidepills <atom.mac@gmail.com>